Prevent unsafe narrowing: ui/views/ for Windows
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3781647
Commit-Queue: Peter Boström <pbos@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027411}

+++ ui/views/widget/desktop_aura/desktop_window_tree_host_win.cc
@@ -378,7 +378,7 @@ void DesktopWindowTreeHostWin::SetShape(
       scaled_rect.roundOut(&rounded_scaled_rect);
       sk_rects.push_back(rounded_scaled_rect);
     }
-    shape.setRects(&sk_rects[0], sk_rects.size());
+    shape.setRects(&sk_rects[0], static_cast<int>(sk_rects.size()));
   } else {
     for (const gfx::Rect& rect : *native_shape)
       shape.op(gfx::RectToSkIRect(rect), SkRegion::kUnion_Op);
