+++ ui/views/widget/widget_hwnd_utils.cc
@@ -61,13 +61,13 @@ void CalculateWindowStylesFromInitParams(
       //   WS_THICKFRAME | WS_MINIMIZEBOX | WS_MAXIMIZEBOX
       *style |= WS_OVERLAPPEDWINDOW;
       if (!widget_delegate->CanMaximize())
-        *style &= ~WS_MAXIMIZEBOX;
+        *style &= static_cast<DWORD>(~WS_MAXIMIZEBOX);
       if (!widget_delegate->CanMinimize())
-        *style &= ~WS_MINIMIZEBOX;
+        *style &= static_cast<DWORD>(~WS_MINIMIZEBOX);
       if (!widget_delegate->CanResize())
-        *style &= ~(WS_THICKFRAME | WS_MAXIMIZEBOX);
+        *style &= static_cast<DWORD>(~(WS_THICKFRAME | WS_MAXIMIZEBOX));
       if (params.remove_standard_frame)
-        *style &= ~(WS_MINIMIZEBOX | WS_MAXIMIZEBOX);
+        *style &= static_cast<DWORD>(~(WS_MINIMIZEBOX | WS_MAXIMIZEBOX));
 
       if (native_widget_delegate->IsDialogBox()) {
         *style |= DS_MODALFRAME;
@@ -87,7 +87,7 @@ void CalculateWindowStylesFromInitParams(
 
       // See layered window comment below.
       if (is_translucent)
-        *style &= ~(WS_THICKFRAME | WS_CAPTION);
+        *style &= static_cast<DWORD>(~(WS_THICKFRAME | WS_CAPTION));
       break;
     }
     case Widget::InitParams::TYPE_CONTROL:
